for #2567, split PaginationEngine
