+++ sharding-core/sharding-core-optimize/src/main/java/org/apache/shardingsphere/core/optimize/statement/sharding/dml/select/pagination/LimitPaginationEngine.java
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.core.optimize.statement.sharding.dml.select.pagination;
+
+import org.apache.shardingsphere.core.parse.sql.segment.dml.pagination.limit.LimitSegment;
+
+import java.util.List;
+
+/**
+ * Pagination engine for limit.
+ *
+ * @author zhangliang
+ */
+public final class LimitPaginationEngine {
+    
+    /**
+     * Create pagination.
+     * 
+     * @param limitSegment limit segment
+     * @param parameters SQL parameters
+     * @return pagination
+     */
+    public Pagination createPagination(final LimitSegment limitSegment, final List<Object> parameters) {
+        return new Pagination(limitSegment.getOffset().orNull(), limitSegment.getRowCount().orNull(), parameters);
+    }
+}
